size_t instead of unsigned int

+++ src/gdxviewer/columnfilterframe.cpp
@@ -71,7 +71,7 @@ void ColumnFilterFrame::apply()
     bool checked;
     std::vector<bool> filterActive = mSymbol->filterActive();
     filterActive[mColumn] = false;
-    for (unsigned int idx=0; idx<uelsInColumn->size(); idx++)
+    for (size_t idx=0; idx<uelsInColumn->size(); idx++)
     {
         checked = mModel->checked()[idx];
         showUelInColumn[uelsInColumn->at(idx)] = checked;
